backend/booking_service/requirements.txt
fastapi
uvicorn
motor
python-jose[cryptography]
pydantic
pydantic-settings
httpx
